requirements.txt
et_xmlfile==2.0.0
lxml==6.0.0
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
